Move NotifySlack func into own package directory
